+++ src/pages/WeatherPage/WeatherPage.jsx
@@ -4,11 +4,10 @@ import './WeatherPage.css';
 import Aside from '../../components/Aside/Aside';
 import { CoordsContext } from '../../context/CoordsProvider';
 import { getLocalCoords, getWeather } from '../../reducers/weather.actions';
-import SearchBar from '../../components/SearchBar/SearchBar';
 
 const WeatherPage = () => {
   console.log('Rendering WeatherPage');
-  // const { getLocalCoords } = useCoords();
+
   const { state, dispatch } = useContext(CoordsContext);
   const { coords, weather } = state;
